resolve unit test file naming issue
